Move `unfill` and `refill` to their own module
